Sort features only once per split
I ran the flamegraph profiler and the `sort_two_vectors` was showing up
as one of the functions that was taking a lot of time.

Under inspection we were sorting every feature with the criteria of
the best feature. But the permutation is always the same no need to
redecide the permutation.

Luckly there is a crate `permutation` that does exactly this. So I used
it and saw a 30% improvement in training time.